+++ hdf/test/sdnmms.c
@@ -77,9 +77,9 @@ test_sdnmms(void)
           for (j = 0; j < 10; j++)
             {
 
-                f64[i][j] = (i * 40) + j;   /* range: 0 ~ 4-billion */
+                f64[i][j] = (float64)((i * 40) + j);   /* range: 0 ~ 4-billion */
 
-                f32[i][j] = (i * 40) + j;   /* range: 0 ~ 4-billion */
+                f32[i][j] = (float32)((i * 40) + j);   /* range: 0 ~ 4-billion */
 
                 i8[i][j] = (int8) ((i * 10) + j);   /* range: 0 ~ 100 */
                 ui8[i][j] = (uint8) ((i * 20) + j);     /* range: 0 ~ 200 */
